findbugs: fix problems
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@50217 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/kernel/cas/AlgoSurdText.java
@@ -514,7 +514,7 @@ public class AlgoSurdText extends AlgoElement implements UsesCAS {
 
 		int numOfAllTerms = vars.length;
 		if (numOfAllTerms - 1 > Math.floor((coeffs.length - 1 - step - offset)
-				/ step)) { // checksum
+				/ (double) step)) { // checksum
 			// appendUndefined();
 			return;
 		}
